src/requirements.txt
# Azure Functions Worker
azure-functions>=1.23.0

# Azure SDK packages for blob storage
azurefunctions-extensions-bindings-blob>=1.1.1

# Durable Functions
azure-functions-durable>=1.2.9

# Azure AI Document Translation
azure-ai-translation-document>=1.0.0

# Azure Blob Storage SDK
azure-storage-blob>=12.19.0

# Image processing
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0
